实现回复接口

+++ src/main/java/life/liudong/community/controller/PublishController.java
@@ -29,7 +29,7 @@ public class PublishController {
             @RequestParam(value = "title" ,required = false) String title,
             @RequestParam(value = "description",required = false) String description,
             @RequestParam(value = "tag",required = false) String tag,
-            @RequestParam(value = "id",required = false) Integer id,
+            @RequestParam(value = "id",required = false) Long id,
             HttpServletRequest request,
             Model model
     ){
@@ -69,7 +69,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,Model model)
+    public String edit(@PathVariable(name = "id") Long id,Model model)
     {
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
