+++ src/main/java/life/liudong/community/controller/QuestionController.java
@@ -16,7 +16,7 @@ public class QuestionController {
 
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name = "id") Integer id, Model model){
+    public String question(@PathVariable(name = "id") Long id, Model model){
 
         QuestionDTO questionDTO=questionService.getById(id);
         //增加阅读
